Improved loadingbar function
If-else statement improved:
At first there was no clause if run outside Jupyter Notebook but with narrow character, where no correction was needed, too.
Now it checks for the exact condition when spaces have to be appended, which is when run outside of Jupyter Notebook with a wide character. If so, append two whitespaces. Else, if for example run inside Jupyter Notebook, which handles wide characters correctly, or if the character is not wide, use the normal definition.